style(link): keep mod.rs under the 2000-line cap
Move the feature-forwarding cargo-args logic into native_features.rs
(apply_native_library_override) and extract the inline
optional_framework_dir_tests mod to a sibling file — the forwarding
block had pushed link/mod.rs from 1996 to 2024 lines.
